添加 molang 补齐功能 - 修正删除作者时，会把模型包也删除的 bug

+++ src/vue/import/new_author.vue
@@ -70,9 +70,11 @@ export default {
                 if (button === 0 && this.isEditAction) {
                     let deleteAuthors = this.authors.splice(this.authorIndex, 1);
                     // 查看头像是否存在，移动到回收站
-                    let avatarPath = join(this.packDirectory, deleteAuthors[0]["avatar"]);
-                    if (fs.existsSync(avatarPath)) {
-                        electron.shell.trashItem(avatarPath);
+                    if (deleteAuthors[0]["avatar"]) {
+                        let avatarPath = join(this.packDirectory, deleteAuthors[0]["avatar"]);
+                        if (fs.existsSync(avatarPath)) {
+                            electron.shell.trashItem(avatarPath);
+                        }
                     }
                     this.newAuthorDialog.close();
                 }
